% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bathy_matrix}
\alias{bathy_matrix}
\title{Example Bathymetric data}
\format{
a \code{matrix} containing \code{numerical} values of bathymetry with deminsion of 1:603 and 1:1151.
}
\usage{
bathy_matrix
}
\description{
An example of bathymetric data as a \code{matrix}
}
\keyword{datasets}
