% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bathy_raster}
\alias{bathy_raster}
\title{Example Bathymetric data as RasterLayer}
\format{
a \code{RasterLayer} containing \code{numerical} values of bathymetry with deminsion of 1:603 and 1:1151.
}
\usage{
bathy_raster
}
\description{
An example of bathymetric data as a \code{RasterLayer}
}
\details{
Used mostly for testing
}
\keyword{datasets}
