% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{build_kdtree_from_pedigree}
\alias{build_kdtree_from_pedigree}
\title{Build kd-tree with all individuals from pedigree}
\usage{
build_kdtree_from_pedigree(pedigree, max_leaf_size)
}
\description{
Build kd-tree with all individuals from pedigree
}
