% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{build_kdtree_from_pids}
\alias{build_kdtree_from_pids}
\title{Build kd-tree from a list of pids}
\usage{
build_kdtree_from_pids(pids, population, max_leaf_size,
  report_progress = FALSE)
}
\description{
Build kd-tree from a list of pids
}
