% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{build_kdtree_from_population}
\alias{build_kdtree_from_population}
\title{Build kd-tree with all individuals from population}
\usage{
build_kdtree_from_population(population, max_leaf_size)
}
\description{
Build kd-tree with all individuals from population
}

