% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{class_transformers}
\alias{class_transformers}
\alias{ac}
\alias{an}
\alias{an.}
\title{Class transformers}
\usage{
ac(x)

an(x)

an.(x)
}
\arguments{
\item{x}{a \code{vector} description}
}
\description{
Class transformers
}
\keyword{internal}
