% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convert_transformers}
\alias{convert_transformers}
\alias{.convert_spr_to_vec}
\title{Convert functions}
\usage{
.convert_spr_to_vec(x)
}
\arguments{
\item{x}{a \code{SpatRaster} object}
}
\description{
Convert an object of one class to another (e.g., \code{SpatRaster} to \code{matrix}
}
\keyword{internal}
