% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_k.R
\name{gaussian_kernal}
\alias{gaussian_kernal}
\title{Gaussian Kernal}
\usage{
gaussian_kernal(size, sigma, muadv = 0, norm = FALSE)
}
\arguments{
\item{size}{a \code{numerical} value that is the size of the areas to calculate the
kernal}

\item{sigma}{a \code{numerical} value that is the standard deviation}

\item{muadv}{a \code{numerical} value that adds to the mean (i.e, mu)}

\item{norm}{a \code{logical} that checks whether to normalize the kernal estimate}
}
\value{
a \code{matrix} that is Gaussian Kernal over an area.
}
\description{
Determines the Gaussian Kernal for an area
}
