% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{index_functions}
\alias{index_functions}
\alias{.valid_idx}
\title{Index functions}
\usage{
.valid_idx(x, na = FALSE)
}
\arguments{
\item{x}{an object to get a index}

\item{na}{a \code{TRUE/FALSE} that controls whether index locations that are \code{NA} are returned. Default is \code{FALSE}.`}
}
\description{
Provides an index number.
}
\keyword{internal}
