% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iniloc}
\alias{iniloc}
\title{Iniloc data}
\format{
\code{data.frame} containing 3 rows and 8 variables
\describe{
\item{fish}{The id of a given tagged fish}
\item{ptt}{The id of a psat tag}
\item{mrpat}{The id of a mrpat tag}
\item{length}{The length of the invidiual}
\item{date}{The date of an event}
\item{lon}{The longitude of an event}
\item{lat}{The latitude of an event}
\item{event}{The event type}
}
}
\usage{
iniloc
}
\description{
An example of psat transmitter info as a \code{data.frame}.
}
\keyword{datasets}
