% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{intergration_helper}
\alias{intergration_helper}
\alias{.integrate_dnorm_rows}
\title{Integrate dnorm}
\usage{
.integrate_dnorm_rows(mean, sd, lower, upper)
}
\arguments{
\item{mean}{the mean value}

\item{sd}{the sd}

\item{lower}{lower bounds}

\item{upper}{upper bounds}
}
\description{
Integrate dnorm
}
\keyword{internal}
