% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_integration.R
\name{likint3_matrix}
\alias{likint3_matrix}
\title{Integrate N(mean, sd) between the minimum and maxium depth for a subset of cells}
\usage{
likint3_matrix(w, widx)
}
\arguments{
\item{w}{a \code{matrix}/\code{data.frame} with columns: mean, sd, lower bound, upper bound}

\item{widx}{a logical vector indicating which rows of \code{w} to integrate - an index}
}
\description{
Vectorized replacement for row-wise stats::integrate(dnorm, ...) --
exact, not an approximation: integral of dnorm from low to high is
pnorm(high) - pnorm(low) by definition of the CDF.
}
\details{
\code{likint3_matrix} use to be called \code{linkint3.sub}
}
\keyword{internal}
