% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{load_functions}
\alias{load_functions}
\alias{load_bathy_raster}
\title{Load functions}
\usage{
load_bathy_raster(bathy)
}
\arguments{
\item{bathy}{A bathymetric feature to be brought in as a \code{SpatRaster} object}
}
\description{
These functions load different data into R
}
\details{
returns a \code{SpatRaster} object to be used with an attribute of index. To access
index use \code{attr(obj, "idx")}.
}
\seealso{
\code{\link[terra:rast]{terra::rast()}}
}
