% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{load_individuals}
\alias{load_individuals}
\title{Construct a population}
\usage{
load_individuals(pid, is_male, pid_mom, pid_dad, is_alive = as.logical(c()),
  birth_year = as.integer(c()), etrs89e = as.numeric(c()),
  etrs89n = as.numeric(c()), progress = TRUE,
  error_on_gender_mismatch = TRUE)
}
\description{
Construct a population
}
