% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_functions.R
\name{make_function}
\alias{make_function}
\alias{make_summary_plot}
\alias{make_summary_table}
\title{Make functions}
\usage{
make_summary_plot(raw_log, output_dir = NULL)

make_summary_table(raw_log, time_res, bin = NULL)
}
\arguments{
\item{raw_log}{a \code{data.frame} containing raw psat logs for an individual.}

\item{output_dir}{a \code{character} string containing the path to save plot or table.}

\item{time_res}{a \code{numeric} that is the median time delay between depth logs, in seconds}

\item{bin}{a \code{numeric} containing the number of seconds within an aggregation (e.g., 3600 seconds is 1 hour)}
}
\description{
These functions make different objects that can be exported like plots and summary tables.
}
