% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_k.R
\name{mask_k}
\alias{mask_k}
\title{Mask k}
\usage{
mask_k(k)
}
\arguments{
\item{k}{is \code{matrix} continaing coordinates}
}
\value{
k as a \code{matrix} that are elucdian distances between locations with maks
}
\description{
This fucntion takes coordates and calcualtes eludian distance between
locations
}
