% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kd.R
\name{meioses_distribution}
\alias{meioses_distribution}
\title{Get distribution of number of meioses within a given radius. 
The Euclidian distance is used based on ETRS89 coordinates due to speed. 
Hence, be sure that this is an acceptable error in comparison to e.g. great-circle distance).}
\usage{
meioses_distribution(population, search_tree_unrelated,
  max_leaf_size_pedigree = 10, pids, radii)
}
\arguments{
\item{population}{A collection of individuals (\code{popr_population})}

\item{search_tree_unrelated}{A search kd-tree when counting unrelated individuals (typically based on \code{population}), created with e.g. \link{\code{build_kdtree_from_population}} or \link{\code{build_kdtree_from_pids}}}

\item{max_leaf_size_pedigree}{A parameter for building kd-trees for the pedigrees that he individuals with pid in \code{pids} belongs to}

\item{pids}{Vector of pids to get meioses_distributions for (must be individuals in \code{population})}

\item{radii}{Vector of radii (radiuses) to search within [km]}
}
\value{
A tibble with meioses distribution information
}
\description{
For each individual (pid) in pids, a search is performed for all individuals in the population within a certain radius.
All the individuals within the radius are classified as either related or unrelated
(depending on whether they belong to the same pedigree or not). If they do belong to the same pedigree,
the number of meioses are found.
}
\details{
Pedigree information is linked to the individuals in the population.
}
