% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kd.R
\name{meioses_distribution_for_tree_pedigrees}
\alias{meioses_distribution_for_tree_pedigrees}
\title{Get distribution of number of meioses within a given radius}
\usage{
meioses_distribution_for_tree_pedigrees(population, pids, radii,
  confirm_all_pedigrees_are_trees = FALSE)
}
\arguments{
\item{population}{A collection of individuals (\code{"popr_population"})}

\item{pids}{Vector of pids to get meioses_distributions for}

\item{radii}{Vector of radii (radiuses) to search within}

\item{confirm_all_pedigrees_are_trees}{Set to TRUE to confirm that all pedigrees are tree (this is the caller's responsibility to ensure that as it would be too time consuming to check this requirement in this function)}
}
\value{
A tibble with meioses distribution information
}
\description{
For each individual (pid) in pids, a search is performed for all individuals in the population within a certain radius. 
All the individuals within the radius are classified as either related or unrelated 
(depending on whether they belong to the same pedigree or not). If they do belong to the same pedigree,
the number of meioses are found.
 
Pedigree information is linked to the individuals in the population. All pedigrees must be trees (FIXME).
}

