% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pdt}
\alias{pdt}
\title{An example of psat transmitter as a \code{data.frame}.}
\format{
\code{data.frame} containing 19,462 rows and 17 variables
\describe{
\item{fish_id}{The id of a given tagged fish}
\item{ptt}{The id of a psat tag}
\item{depth_sensor}{The type of sensor e.g., (\code{0.5})}
\item{source}{Transmission source}
\item{instr}{The intstrument type}
\item{date}{The date recored}
\item{time}{The tiem recorded}
\item{date_time}{The date timestamp}
\item{location_quality}{The quality of the location}
\item{latitude}{The latitude of the transmision}
\item{depth}{The depth of the transmitter}
\item{d_range}{The depth range}
\item{temperature}{The temperature of the the transmitter}
\item{t_range}{The temeprature range}
\item{activity}{An activity measurement}
\item{a_range}{An activity range}
}
}
\usage{
pdt
}
\description{
An example of psat transmitter as a \code{data.frame}.
}
\keyword{datasets}
