% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{safe_functions}
\alias{safe_functions}
\alias{.safe_max}
\alias{.safe_range_val}
\title{Safe base functions}
\usage{
.safe_max(x)

.safe_range_val(x, type = c("min", "max"))
}
\arguments{
\item{x}{a \code{numerical} vector}

\item{type}{a \code{character} determining whether \code{min()} or \code{max()} is used}
}
\description{
These functions check if x is \code{NA} then
return a stat (e.g, range, min, max)
}
\keyword{internal}
