% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_likelihood.R
\name{calc_likdepth_transmit}
\alias{calc_likdepth_transmit}
\title{Calculate Likelihood Depth Transmitted}
\usage{
calc_likdepth_transmit(preped_log, preped_bathy, daily_sd_depth = 0.57)
}
\arguments{
\item{preped_log}{a \code{list} containing summary objects crated by \code{prep_summary_likdepth()}.}

\item{preped_bathy}{a \code{list} containinging objects created by \code{prep_bathy_likdepth()}.}

\item{daily_sd_depth}{a numeric value}
}
\description{
These are functions that will calculated the likelihood of depth occupancy based on
depth values transmitted and bathymetry
}
