% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_functions.R
\name{error_functions}
\alias{error_functions}
\alias{error_empty_file}
\alias{error_load}
\alias{error_path}
\alias{error_raster}
\title{Error functions}
\usage{
error_empty_file(path)

error_load(path)

error_path(path)

error_raster(raster)
}
\arguments{
\item{path}{The file path to a file.}

\item{raster}{a raster object to assess}
}
\description{
These functions produce error message displayed to the user.
}
\keyword{internal}
