% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{index_functions}
\alias{index_functions}
\alias{.valid_idx}
\title{Index functions}
\usage{
.valid_idx(x)
}
\arguments{
\item{x}{an object to get a index}
}
\description{
Provides an index number.
}
\keyword{internal}
