% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_integration.R
\name{likelihood_integration}
\alias{likelihood_integration}
\alias{likint3}
\title{Likelihood integration functions}
\usage{
likint3(w, wsd, minT, maxT)
}
\arguments{
\item{w}{a \code{vector} description}

\item{wsd}{a \code{vector} containging the standard deviation}

\item{minT}{a \code{vector} containing the minmum temperature}

\item{maxT}{a \code{vector} containing the maximum temperature}
}
\description{
Likelihood integration functions
}
\keyword{internal}
