% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgs_functions.R
\name{msg_function}
\alias{msg_function}
\alias{msg_bathy}
\alias{msg_dates}
\alias{msg_load_raster}
\alias{msg_load_rdata}
\alias{msg_start}
\title{Message functions}
\usage{
msg_bathy(bathy)

msg_dates(df)

msg_load_raster(path)

msg_load_rdata(obj, obj_name)

msg_start(chr, type, t0 = round(Sys.time()))
}
\arguments{
\item{bathy}{a bathymetric feature.}

\item{df}{a \code{data.frame}}

\item{path}{a file path. `}

\item{obj}{an object}

\item{obj_name}{the name of an object}

\item{chr}{a \code{character} string that tells which likelihood such as e.g., \code{"Depth"}}

\item{type}{a \code{character} string that tells whehter it is likelhood or summary.}

\item{t0}{current time}
}
\description{
These functions produce information to the user.
}
\keyword{internal}
