% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgs_functions.R
\name{msg_function}
\alias{msg_function}
\alias{msg_bathy}
\alias{msg_dates}
\alias{msg_end}
\alias{msg_load_raster}
\alias{msg_load_rdata}
\alias{msg_start}
\title{Message functions}
\usage{
msg_bathy(bathy)

msg_dates(df)

msg_end(chr, t1 = round(Sys.time()), t0)

msg_load_raster(path)

msg_load_rdata(obj, obj_name)

msg_start(chr, t0 = round(Sys.time()))
}
\arguments{
\item{bathy}{a bathymetric feature.}

\item{df}{a \code{data.frame}}

\item{chr}{a \code{character} string that tells which likelihood such as e.g., \code{"Depth"}}

\item{t1}{end time}

\item{t0}{current time}

\item{path}{a file path.}

\item{obj}{an object}

\item{obj_name}{the name of an object}
}
\description{
These functions produce information to the user.
}
\keyword{internal}
