% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_functions.R
\name{prep_summary_likdepth}
\alias{prep_summary_likdepth}
\title{Prepare Summary Data}
\usage{
prep_summary_likdepth(
  raw_log,
  iniloc,
  bin = bin,
  output_dir = NULL,
  make_plot = TRUE
)
}
\arguments{
\item{raw_log}{a \code{data.frame} containing the raw pop-off satalitte tag log}

\item{iniloc}{a \code{data.frame} containing the iniloc information about when the tag was deployed, mrpat, and pop dates and location.}

\item{bin}{a \code{numeric} containing the number of seconds within an aggregation (e.g., 3600 seconds is 1 hour).}

\item{output_dir}{Defaults to \code{NULL} and will not export plot. Please provide a \code{character} string containing the path deseired
to save the plot.}

\item{make_plot}{a \code{logical} value that controls whether the to make the summary plot. Default is \code{TRUE}.}
}
\description{
This function prepared summary data to be used in the likelihood estimates.
}
